Consolidate exercise question list reports. Add missing resource handling.
